requirements-audio.txt
# REQUIREMENTS-AUDIO.TXT - Audio Processing Dependencies
numpy==1.26.4

# AUDIO PROCESSING & TRANSCRIPTION
faster-whisper==1.0.3
ctranslate2==4.6.0
tokenizers==0.19.1

# Audio manipulation
pydub==0.25.1
soundfile==0.13.1
av==12.3.0

# Audio analysis
librosa==0.10.2.post1
audioread==3.0.1
lazy-loader==0.4
msgpack==1.1.0
pooch==1.8.2
soxr==0.5.0.post1

# SPEAKER DIARIZATION 
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyannoteai-sdk==0.3.0

# Pyannote dependencies
asteroid-filterbanks==0.4.0
einops==0.8.1
pytorch-metric-learning==2.9.0
semver==3.0.4

# ML TRAINING & OPTIMIZATION
pytorch-lightning==2.5.5
lightning==2.5.5
lightning-utilities==0.15.2
tensorboardX==2.6.4
torchmetrics==1.8.2

# Transformers
transformers==4.40.2
sentencepiece==0.2.1
safetensors==0.6.2
accelerate==1.11.0

# Audio ML
speechbrain==1.0.3
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
julius==0.2.7

# PDF & DOCUMENT GENERATION
fpdf==1.7.2
reportlab==4.2.5
markdown==3.7
huggingface_hub==0.36.0

# LLM & AI APIs 
openai==1.59.5
anthropic==0.42.0
litellm==1.56.4

# DATA PROCESSING
pandas==2.3.3
scikit-learn==1.7.2
joblib==1.5.2
threadpoolctl==3.6.0

# VISUALIZATION & MEDIA
matplotlib==3.10.7
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.2.1
proglog==0.1.12
decorator==5.2.1

# Matplotlib dependencies
contourpy==1.3.2
cycler==0.12.1
fonttools==4.60.1
kiwisolver==1.4.9

# CONFIGURATION & UTILITIES
omegaconf==2.3.0
HyperPyYAML==1.2.2
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.14

# CLI & Logging
coloredlogs==15.0.1
colorlog==6.10.1
humanfriendly==10.0
markdown-it-py==4.0.0
mdurl==0.1.2

# File handling
fsspec==2024.9.0


# OPTIMIZATION
optuna==4.5.0

# TELEMETRY
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-semantic-conventions==0.59b0
googleapis-common-protos==1.71.0
grpcio==1.76.0

# UTILITIES
charset-normalizer==3.4.4
regex==2025.11.3
networkx==3.4.2
sympy==1.14.0
mpmath==1.3.0
tabulate==0.9.0
wrapt==2.0.0
flatbuffers==25.9.23
onnxruntime==1.23.2
annotated-doc==0.0.3
antlr4-python3-runtime==4.9.3
marisa-trie==1.3.1
torchcodec==0.8.1
docopt==0.6.2
primePy==1.3
sortedcontainers==2.4.0
pyreadline3==3.5.4
pyparsing==3.1.4
packaging==25.0
python-dateutil==2.9.0.post0
scipy==1.14.1
rich==13.7.1